+++ tests/baselines/reference/unicodeEscapesInNames01(target=es2015).types
@@ -1,4 +1,40 @@
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithEscape1.ts ===
+=== tests/cases/compiler/identifierVariableWithEscape1.ts ===
+export let \u0078 = 10;
+>\u0078 : number
+>10 : 10
+
+x++;
+>x++ : number
+>x : number
+
+=== tests/cases/compiler/identifierVariableWithEscape2.ts ===
+export let x\u0078 = 10;
+>x\u0078 : number
+>10 : 10
+
+xx++;
+>xx++ : number
+>xx : number
+
+=== tests/cases/compiler/identifierVariableWithExtendedEscape1.ts ===
+export let \u{78} = 10;
+>\u{78} : number
+>10 : 10
+
+x++;
+>x++ : number
+>x : number
+
+=== tests/cases/compiler/identifierVariableWithExtendedEscape2.ts ===
+export let x\u{78} = 10;
+>x\u{78} : number
+>10 : 10
+
+xx++;
+>xx++ : number
+>xx : number
+
+=== tests/cases/compiler/IdentifierNameWithEscape1.ts ===
 export class IdentifierNameWithEscape1 {
 >IdentifierNameWithEscape1 : IdentifierNameWithEscape1
 
@@ -26,7 +62,7 @@ export class IdentifierNameWithEscape1 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithEscape2.ts ===
+=== tests/cases/compiler/IdentifierNameWithEscape2.ts ===
 export class IdentifierNameWithEscape2 {
 >IdentifierNameWithEscape2 : IdentifierNameWithEscape2
 
@@ -54,7 +90,7 @@ export class IdentifierNameWithEscape2 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithExtendedEscape1.ts ===
+=== tests/cases/compiler/IdentifierNameWithExtendedEscape1.ts ===
 export class IdentifierNameWithExtendedEscape1 {
 >IdentifierNameWithExtendedEscape1 : IdentifierNameWithExtendedEscape1
 
@@ -82,7 +118,7 @@ export class IdentifierNameWithExtendedEscape1 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithExtendedEscape2.ts ===
+=== tests/cases/compiler/IdentifierNameWithExtendedEscape2.ts ===
 export class IdentifierNameWithExtendedEscape2 {
 >IdentifierNameWithExtendedEscape2 : IdentifierNameWithExtendedEscape2
 
@@ -110,7 +146,7 @@ export class IdentifierNameWithExtendedEscape2 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithEscape1.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithEscape1.ts ===
 export class PrivateIdentifierWithEscape1 {
 >PrivateIdentifierWithEscape1 : PrivateIdentifierWithEscape1
 
@@ -136,7 +172,7 @@ export class PrivateIdentifierWithEscape1 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithEscape2.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithEscape2.ts ===
 export class PrivateIdentifierWithEscape2 {
 >PrivateIdentifierWithEscape2 : PrivateIdentifierWithEscape2
 
@@ -162,7 +198,7 @@ export class PrivateIdentifierWithEscape2 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithExtendedEscape1.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape1.ts ===
 export class PrivateIdentifierWithExtendedEscape1 {
 >PrivateIdentifierWithExtendedEscape1 : PrivateIdentifierWithExtendedEscape1
 
@@ -188,7 +224,7 @@ export class PrivateIdentifierWithExtendedEscape1 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithExtendedEscape2.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape2.ts ===
 export class PrivateIdentifierWithExtendedEscape2 {
 >PrivateIdentifierWithExtendedEscape2 : PrivateIdentifierWithExtendedEscape2
