+++ tests/baselines/reference/unicodeEscapesInNames01(target=es5).errors.txt
@@ -1,0 +1,134 @@
+tests/cases/compiler/PrivateIdentifierNameWithEscape1.ts(2,5): error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+tests/cases/compiler/PrivateIdentifierNameWithEscape2.ts(2,5): error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape1.ts(2,5): error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape2.ts(2,5): error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+
+
+==== tests/cases/compiler/identifierVariableWithEscape1.ts (0 errors) ====
+    export let \u0078 = 10;
+    x++;
+    
+==== tests/cases/compiler/identifierVariableWithEscape2.ts (0 errors) ====
+    export let x\u0078 = 10;
+    xx++;
+    
+==== tests/cases/compiler/identifierVariableWithExtendedEscape1.ts (0 errors) ====
+    export let \u{78} = 10;
+    x++;
+    
+==== tests/cases/compiler/identifierVariableWithExtendedEscape2.ts (0 errors) ====
+    export let x\u{78} = 10;
+    xx++;
+    
+==== tests/cases/compiler/IdentifierNameWithEscape1.ts (0 errors) ====
+    export class IdentifierNameWithEscape1 {
+        \u0078: number;
+    
+        constructor() {
+            this.\u0078 = 0;
+        }
+    
+        doThing() {
+            this.x = 42;
+        }
+    }
+    
+==== tests/cases/compiler/IdentifierNameWithEscape2.ts (0 errors) ====
+    export class IdentifierNameWithEscape2 {
+        x\u0078: number;
+    
+        constructor() {
+            this.x\u0078 = 0;
+        }
+    
+        doThing() {
+            this.xx = 42;
+        }
+    }
+    
+==== tests/cases/compiler/IdentifierNameWithExtendedEscape1.ts (0 errors) ====
+    export class IdentifierNameWithExtendedEscape1 {
+        \u{78}: number;
+    
+        constructor() {
+            this.\u{78} = 0;
+        }
+    
+        doThing() {
+            this.x = 42;
+        }
+    }
+    
+==== tests/cases/compiler/IdentifierNameWithExtendedEscape2.ts (0 errors) ====
+    export class IdentifierNameWithExtendedEscape2 {
+        x\u{78}: number;
+    
+        constructor() {
+            this.x\u{78} = 0;
+        }
+    
+        doThing() {
+            this.xx = 42;
+        }
+    }
+    
+==== tests/cases/compiler/PrivateIdentifierNameWithEscape1.ts (1 errors) ====
+    export class PrivateIdentifierWithEscape1 {
+        #\u0078: number;
+        ~~~~~~~
+!!! error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+    
+        constructor() {
+            this.#\u0078 = 0;
+        }
+    
+        doThing() {
+            this.#x = 42;
+        }
+    }
+    
+==== tests/cases/compiler/PrivateIdentifierNameWithEscape2.ts (1 errors) ====
+    export class PrivateIdentifierWithEscape2 {
+        #x\u0078: number;
+        ~~~~~~~~
+!!! error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+    
+        constructor() {
+            this.#x\u0078 = 0;
+        }
+    
+        doThing() {
+            this.#xx = 42;
+        }
+    }
+    
+==== tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape1.ts (1 errors) ====
+    export class PrivateIdentifierWithExtendedEscape1 {
+        #\u{78}: number;
+        ~~~~~~~
+!!! error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+    
+        constructor() {
+            this.#\u{78} = 0;
+        }
+    
+        doThing() {
+            this.#x = 42;
+        }
+    }
+    
+==== tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape2.ts (1 errors) ====
+    export class PrivateIdentifierWithExtendedEscape2 {
+        #x\u{78}: number;
+        ~~~~~~~~
+!!! error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
+    
+        constructor() {
+            this.#x\u{78} = 0;
+        }
+    
+        doThing() {
+            this.#xx = 42;
+        }
+    }
+
