+++ tests/baselines/reference/unicodeEscapesInNames01(target=es5).types
@@ -1,4 +1,40 @@
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithEscape1.ts ===
+=== tests/cases/compiler/identifierVariableWithEscape1.ts ===
+export let \u0078 = 10;
+>\u0078 : number
+>10 : 10
+
+x++;
+>x++ : number
+>x : number
+
+=== tests/cases/compiler/identifierVariableWithEscape2.ts ===
+export let x\u0078 = 10;
+>x\u0078 : number
+>10 : 10
+
+xx++;
+>xx++ : number
+>xx : number
+
+=== tests/cases/compiler/identifierVariableWithExtendedEscape1.ts ===
+export let \u{78} = 10;
+>\u{78} : number
+>10 : 10
+
+x++;
+>x++ : number
+>x : number
+
+=== tests/cases/compiler/identifierVariableWithExtendedEscape2.ts ===
+export let x\u{78} = 10;
+>x\u{78} : number
+>10 : 10
+
+xx++;
+>xx++ : number
+>xx : number
+
+=== tests/cases/compiler/IdentifierNameWithEscape1.ts ===
 export class IdentifierNameWithEscape1 {
 >IdentifierNameWithEscape1 : IdentifierNameWithEscape1
 
@@ -26,7 +62,7 @@ export class IdentifierNameWithEscape1 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithEscape2.ts ===
+=== tests/cases/compiler/IdentifierNameWithEscape2.ts ===
 export class IdentifierNameWithEscape2 {
 >IdentifierNameWithEscape2 : IdentifierNameWithEscape2
 
@@ -54,78 +90,63 @@ export class IdentifierNameWithEscape2 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithExtendedEscape1.ts ===
+=== tests/cases/compiler/IdentifierNameWithExtendedEscape1.ts ===
 export class IdentifierNameWithExtendedEscape1 {
 >IdentifierNameWithExtendedEscape1 : IdentifierNameWithExtendedEscape1
 
     \u{78}: number;
->u : any
->78 : 78
->number : any
+>\u{78} : number
 
     constructor() {
->constructor() : any
->constructor : any
-
         this.\u{78} = 0;
->this. : any
->this : undefined
-> : any
->u : any
->78 : 78
+>this.\u{78} = 0 : 0
+>this.\u{78} : number
+>this : this
+>\u{78} : number
 >0 : 0
     }
 
     doThing() {
->doThing() : any
->doThing : any
+>doThing : () => void
 
         this.x = 42;
 >this.x = 42 : 42
->this.x : any
->this : undefined
->x : any
+>this.x : number
+>this : this
+>x : number
 >42 : 42
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/IdentifierNameWithExtendedEscape2.ts ===
+=== tests/cases/compiler/IdentifierNameWithExtendedEscape2.ts ===
 export class IdentifierNameWithExtendedEscape2 {
 >IdentifierNameWithExtendedEscape2 : IdentifierNameWithExtendedEscape2
 
     x\u{78}: number;
->x : any
->u : any
->78 : 78
->number : any
+>x\u{78} : number
 
     constructor() {
->constructor() : any
->constructor : any
-
         this.x\u{78} = 0;
->this.x : any
->this : undefined
->x : any
->u : any
->78 : 78
+>this.x\u{78} = 0 : 0
+>this.x\u{78} : number
+>this : this
+>x\u{78} : number
 >0 : 0
     }
 
     doThing() {
->doThing() : any
->doThing : any
+>doThing : () => void
 
         this.xx = 42;
 >this.xx = 42 : 42
->this.xx : any
->this : undefined
->xx : any
+>this.xx : number
+>this : this
+>xx : number
 >42 : 42
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithEscape1.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithEscape1.ts ===
 export class PrivateIdentifierWithEscape1 {
 >PrivateIdentifierWithEscape1 : PrivateIdentifierWithEscape1
 
@@ -151,7 +172,7 @@ export class PrivateIdentifierWithEscape1 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithEscape2.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithEscape2.ts ===
 export class PrivateIdentifierWithEscape2 {
 >PrivateIdentifierWithEscape2 : PrivateIdentifierWithEscape2
 
@@ -177,70 +198,54 @@ export class PrivateIdentifierWithEscape2 {
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithExtendedEscape1.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape1.ts ===
 export class PrivateIdentifierWithExtendedEscape1 {
 >PrivateIdentifierWithExtendedEscape1 : PrivateIdentifierWithExtendedEscape1
 
     #\u{78}: number;
-># : any
->u : any
->78 : 78
->number : any
+>#\u{78} : number
 
     constructor() {
->constructor() : any
->constructor : any
-
         this.#\u{78} = 0;
->this.# : any
->this : undefined
->u : any
->78 : 78
+>this.#\u{78} = 0 : 0
+>this.#\u{78} : number
+>this : this
 >0 : 0
     }
 
     doThing() {
->doThing() : any
->doThing : any
+>doThing : () => void
 
         this.#x = 42;
 >this.#x = 42 : 42
->this.#x : any
->this : undefined
+>this.#x : number
+>this : this
 >42 : 42
     }
 }
 
-=== tests/cases/conformance/classes/members/privateNames/PrivateIdentifierNameWithExtendedEscape2.ts ===
+=== tests/cases/compiler/PrivateIdentifierNameWithExtendedEscape2.ts ===
 export class PrivateIdentifierWithExtendedEscape2 {
 >PrivateIdentifierWithExtendedEscape2 : PrivateIdentifierWithExtendedEscape2
 
     #x\u{78}: number;
->#x : any
->u : any
->78 : 78
->number : any
+>#x\u{78} : number
 
     constructor() {
->constructor() : any
->constructor : any
-
         this.#x\u{78} = 0;
->this.#x : any
->this : undefined
->u : any
->78 : 78
+>this.#x\u{78} = 0 : 0
+>this.#x\u{78} : number
+>this : this
 >0 : 0
     }
 
     doThing() {
->doThing() : any
->doThing : any
+>doThing : () => void
 
         this.#xx = 42;
 >this.#xx = 42 : 42
->this.#xx : any
->this : undefined
+>this.#xx : number
+>this : this
 >42 : 42
     }
 }
