+++ tests/baselines/reference/unicodeEscapesInNames02(target=es2015).symbols
@@ -1,0 +1,65 @@
+=== tests/cases/compiler/extendedEscapesForAstralsInVarsAndClasses.ts ===
+// Example from https://mathiasbynens.be/notes/javascript-identifiers-es6
+// Astral characters should be accepted in ES2015
+
+// U+102A7 CARIAN LETTER A2
+var 𐊧: string;
+>𐊧 : Symbol(𐊧, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 4, 3), Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 3))
+
+var \u{102A7}: string;
+>\u{102A7} : Symbol(𐊧, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 4, 3), Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 3))
+
+if (Math.random()) {
+>Math.random : Symbol(Math.random, Decl(lib.es5.d.ts, --, --))
+>Math : Symbol(Math, Decl(lib.es5.d.ts, --, --), Decl(lib.es5.d.ts, --, --), Decl(lib.es2015.core.d.ts, --, --), Decl(lib.es2015.symbol.wellknown.d.ts, --, --))
+>random : Symbol(Math.random, Decl(lib.es5.d.ts, --, --))
+
+    𐊧 = "hello";
+>𐊧 : Symbol(𐊧, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 4, 3), Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 3))
+}
+else {
+    \u{102A7} = "hallo";
+>\u{102A7} : Symbol(𐊧, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 4, 3), Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 3))
+}
+
+class Foo {
+>Foo : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+
+    \u{102A7}: string;
+>\u{102A7} : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+
+    constructor() {
+        this.\u{102A7} = " world";
+>this.\u{102A7} : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+>this : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+>\u{102A7} : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+    }
+    methodA() {
+>methodA : Symbol(Foo.methodA, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 18, 5))
+
+        return this.𐊧;
+>this.𐊧 : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+>this : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+>𐊧 : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+    }
+}
+
+export var _𐊧 = new Foo().\u{102A7} + new Foo().methodA();
+>_𐊧 : Symbol(_𐊧, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 24, 10))
+>new Foo().\u{102A7} : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+>Foo : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+>\u{102A7} : Symbol(Foo[\u{102A7}], Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 14, 11))
+>new Foo().methodA : Symbol(Foo.methodA, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 18, 5))
+>Foo : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+>methodA : Symbol(Foo.methodA, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 18, 5))
+
+_\u{102A7} += "!";
+>_\u{102A7} : Symbol(_𐊧, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 24, 10))
+
+=== tests/cases/compiler/astralAsSurrogatePair.ts ===
+import { _𐊧 as \uD800\uDEA7 } from "./extendedEscapesForAstralsInVarsAndClasses.js";
+>_𐊧 : Symbol((Missing), Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 24, 10))
+> : Symbol((Missing), Decl(astralAsSurrogatePair.ts, 0, 8))
+>uD800 : Symbol(uD800, Decl(astralAsSurrogatePair.ts, 0, 17))
+>uDEA7 : Symbol(uDEA7, Decl(astralAsSurrogatePair.ts, 0, 23))
+
