+++ tests/baselines/reference/unicodeEscapesInNames02(target=es5).symbols
@@ -1,0 +1,62 @@
+=== tests/cases/compiler/extendedEscapesForAstralsInVarsAndClasses.ts ===
+// Example from https://mathiasbynens.be/notes/javascript-identifiers-es6
+// Astral characters should be accepted in ES2015
+
+// U+102A7 CARIAN LETTER A2
+var 𐊧: string;
+>string : Symbol(string, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 4, 7))
+
+var \u{102A7}: string;
+>u : Symbol(u, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 5))
+>A7 : Symbol(A7, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 7))
+
+if (Math.random()) {
+>Math.random : Symbol(Math.random, Decl(lib.es5.d.ts, --, --))
+>Math : Symbol(Math, Decl(lib.es5.d.ts, --, --), Decl(lib.es5.d.ts, --, --))
+>random : Symbol(Math.random, Decl(lib.es5.d.ts, --, --))
+
+    𐊧 = "hello";
+}
+else {
+    \u{102A7} = "hallo";
+>u : Symbol(u, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 5))
+>A7 : Symbol(A7, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 7))
+}
+
+class Foo {
+>Foo : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+
+    \u{102A7}: string;
+>u : Symbol(Foo.u, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 15, 5))
+>A7 : Symbol(A7, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 7))
+>string : Symbol(string, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 4, 7))
+
+    constructor() {
+        this.\u{102A7} = " world";
+>u : Symbol(u, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 5))
+>A7 : Symbol(A7, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 7))
+    }
+    methodA() {
+        return this.𐊧;
+    }
+}
+
+export var _𐊧 = new Foo().\u{102A7} + new Foo().methodA();
+>_ : Symbol(_, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 24, 10))
+>Foo : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+>u : Symbol(u, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 5))
+>A7 : Symbol(A7, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 7))
+>Foo : Symbol(Foo, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 12, 1))
+
+_\u{102A7} += "!";
+>_ : Symbol(_, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 24, 10))
+>u : Symbol(u, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 5))
+>A7 : Symbol(A7, Decl(extendedEscapesForAstralsInVarsAndClasses.ts, 5, 7))
+
+=== tests/cases/compiler/astralAsSurrogatePair.ts ===
+import { _𐊧 as \uD800\uDEA7 } from "./extendedEscapesForAstralsInVarsAndClasses.js";
+>_ : Symbol(_, Decl(astralAsSurrogatePair.ts, 0, 8))
+>as : Symbol(as, Decl(astralAsSurrogatePair.ts, 0, 12))
+>uD800 : Symbol(uD800, Decl(astralAsSurrogatePair.ts, 0, 17))
+>uDEA7 : Symbol(uDEA7, Decl(astralAsSurrogatePair.ts, 0, 23))
+
